Added authenticationManager
