+++ app/build.gradle
@@ -13,7 +13,7 @@ android {
     }
     productFlavors {
         develop {
-            buildConfigField 'String', 'HOST', '"http://192.168.1.132:8080"'
+            buildConfigField 'String', 'HOST', '"http://192.168.0.15:8080"'
             buildConfigField 'String', 'USERNAME', '"TestUser"'
             buildConfigField 'String', 'PASSWORD', '"TestPassword"'
         }
