+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".SplashActivity"
+        <activity android:name=".activities.SplashActivity"
             android:theme="@style/SplashTheme">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -19,12 +19,12 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".LoginActivity"
+            android:name=".activities.LoginActivity"
             android:configChanges="orientation"
             android:screenOrientation="portrait">
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name=".activities.MainActivity"
             android:configChanges="orientation"
             android:screenOrientation="portrait" />
 
@@ -35,7 +35,7 @@
             android:name="com.google.android.geo.API_KEY"
             android:value="AIzaSyD4GZsErcOopCqw7zHZNr-2qLqy6y_gJ_s" />
 
-        <activity android:name=".RegisterActivity"
+        <activity android:name=".activities.RegisterActivity"
             android:configChanges="orientation"
             android:screenOrientation="portrait"/>
