+++ app/src/main/java/com/gut/follower/activities/LoginActivity.java
@@ -1,0 +1,52 @@
+package com.gut.follower.activities;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+import android.view.View;
+import android.widget.Button;
+import android.widget.EditText;
+import android.widget.TextView;
+
+import com.gut.follower.R;
+import com.gut.follower.utility.AuthenticationManager;
+
+public class LoginActivity extends AppCompatActivity {
+
+    private EditText username;
+    private EditText password;
+    private Button loginButton;
+    private TextView register;
+
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_login);
+
+        username = (EditText)findViewById(R.id.username);
+        password = (EditText)findViewById(R.id.password);
+        loginButton = (Button)findViewById(R.id.login_btn);
+        register = (TextView)findViewById(R.id.register);
+
+
+        loginButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                new AuthenticationManager(LoginActivity.this, username, password).login();
+            }
+        });
+
+        register.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                goToRegisterActivity();
+            }
+        });
+    }
+
+    private void goToRegisterActivity() {
+        Intent intent = new Intent(getApplicationContext(), RegisterActivity.class);
+        startActivity(intent);
+    }
+}
