+++ app/src/main/java/com/gut/follower/activities/MainActivity.java
@@ -1,10 +1,13 @@
-package com.gut.follower;
+package com.gut.follower.activities;
 
 import android.support.annotation.IdRes;
 import android.support.v4.app.Fragment;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
+import com.gut.follower.R;
+import com.gut.follower.activities.fragments.HistoryFragment;
+import com.gut.follower.activities.fragments.RecordFragment;
 import com.roughike.bottombar.BottomBar;
 import com.roughike.bottombar.OnMenuTabClickListener;
