+++ app/src/main/java/com/gut/follower/activities/fragments/HistoryFragment.java
@@ -1,4 +1,4 @@
-package com.gut.follower;
+package com.gut.follower.activities.fragments;
 
 
 import android.os.Bundle;
@@ -6,6 +6,8 @@ import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.gut.follower.R;
 
 
 /**
