+++ app/src/main/res/layout/activity_main.xml
@@ -8,6 +8,6 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.gut.follower.MainActivity">
+    tools:context=".activities.MainActivity">
 
 </FrameLayout>
