+++ app/src/main/res/layout/activity_register.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.gut.follower.LoginActivity">
+    tools:context=".activities.LoginActivity">
 
     <ImageView
         android:layout_width="match_parent"
