+++ app/src/main/res/layout/fragment_history.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.gut.follower.HistoryFragment">
+    tools:context=".activities.fragments.HistoryFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
