+++ app/src/main/res/layout/fragment_record.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.gut.follower.RecordFragment">
+    tools:context=".activities.fragments.RecordFragment">
 
     <fragment
         android:id="@+id/map"
